task/move update message to a new file
