+++ components/index.ts
@@ -29,6 +29,7 @@ export { default as DIDOperations } from './SSI/DIDxWallet/Wallet/DIDOperations'
 export { default as NewDoc } from './SSI/DIDxWallet/Wallet/DIDOperations/NewDoc'
 export { default as CommonLinks } from './SSI/DIDxWallet/Wallet/DIDOperations/CommonLinks'
 export { default as SocialCard } from './SSI/DIDxWallet/Wallet/DIDOperations/SocialCard'
+export { default as UpdateMessage } from './SSI/DIDxWallet/Wallet/DIDOperations/UpdateMessage'
 export { default as SubmitCreate } from './SSI/DIDxWallet/Wallet/DIDOperations/SubmitCreate'
 export { default as SubmitRecover } from './SSI/DIDxWallet/Wallet/DIDOperations/SubmitRecover'
 export { default as Liquidity } from './Liquidity'
